requirements.txt
pyautogui
opencv-python
Pillow
numpy
pytesseract
keyboard